또 다른 삭제 방법

+++ crud_practice.py
@@ -3,6 +3,10 @@
 # 2. 파일 전체 삭제하기
 
 # 파일 전체 삭제할 경우 os 도구를 빌려와야 한다.
-#1.
+#2.
 import os
-os.remove("instructor.txt")
+if os.path.exists("instructor.txt"):
+    os.remove("instructor.txt")
+    print("파일이 완전히 삭제되었습니다")
+else:
+    print("삭제할 파일이 없습니다")
